fix: move tests that require L1ChainSpec to 'tests' folder
