+++ crates/edr_provider/Cargo.toml
@@ -38,7 +38,7 @@ tracing = { version = "0.1.37", features = ["attributes", "std"], optional = tru
 
 [dev-dependencies]
 anyhow = "1.0.89"
-edr_chain_l1.workspace = true
+edr_chain_l1 = { workspace = true, features = ["test-utils"] }
 edr_evm = { path = "../edr_evm", features = ["test-utils"] }
 edr_test_utils.workspace = true
 paste = { version = "1.0.14", default-features = false }
