+++ crates/edr_provider/src/requests/eth/transactions.rs
@@ -377,53 +377,3 @@ You can use them by running Hardhat Network with 'hardfork' {minimum_hardfork:?}
         },
     )
 }
-
-#[cfg(test)]
-mod tests {
-    use anyhow::Context;
-    use edr_eth::{Address, Bytes, U256};
-    use transaction::{signed::FakeSign as _, TxKind};
-
-    use super::*;
-    use crate::test_utils::{one_ether, ProviderTestFixture};
-
-    #[test]
-    fn transaction_by_hash_for_impersonated_account() -> anyhow::Result<()> {
-        let mut fixture = ProviderTestFixture::<L1ChainSpec>::new_local()?;
-
-        let impersonated_account: Address = "0x20620fa0ad46516e915029c94e3c87c9cd7861ff".parse()?;
-        fixture
-            .provider_data
-            .impersonate_account(impersonated_account);
-
-        fixture
-            .provider_data
-            .set_balance(impersonated_account, one_ether())?;
-
-        let chain_id = fixture.provider_data.chain_id();
-
-        let transaction = transaction::Request::Eip155(transaction::request::Eip155 {
-            kind: TxKind::Call(Address::ZERO),
-            gas_limit: 30_000,
-            gas_price: 42_000_000_000,
-            value: U256::from(1),
-            input: Bytes::default(),
-            nonce: 0,
-            chain_id,
-        })
-        .fake_sign(impersonated_account);
-        let transaction = transaction::validate(transaction, fixture.provider_data.evm_spec_id())?;
-
-        fixture.provider_data.set_auto_mining(true);
-        let result = fixture.provider_data.send_transaction(transaction)?;
-        assert!(result.transaction_result_and_trace().is_some());
-
-        let rpc_transaction =
-            handle_get_transaction_by_hash(&fixture.provider_data, result.transaction_hash)?
-                .context("transaction not found")?;
-        assert_eq!(&rpc_transaction.from, &impersonated_account);
-        assert_eq!(&rpc_transaction.hash, &result.transaction_hash);
-
-        Ok(())
-    }
-}
